moved staticinfoaccumulator test to a new file
